+++ noodles-csi/src/async/io/writer/index/reference_sequences/bins.rs
@@ -6,7 +6,9 @@ use tokio::io::{self, AsyncWrite, AsyncWriteExt};
 
 use self::chunks::write_chunks;
 use super::write_metadata;
-use crate::binning_index::index::reference_sequence::{Bin, Metadata};
+use crate::binning_index::index::reference_sequence::{
+    Bin, Metadata, index::BinnedIndex, parent_id,
+};
 
 pub(super) async fn write_bins<W>(
     writer: &mut W,
@@ -31,9 +33,9 @@ where
 
     writer.write_i32_le(n_bin).await?;
 
-    for (id, bin) in bins {
-        let first_record_start_position = index.get(id).copied().unwrap_or_default();
-        write_bin(writer, *id, first_record_start_position, bin).await?;
+    for (&id, bin) in bins {
+        let first_record_start_position = first_record_start_position(index, id);
+        write_bin(writer, id, first_record_start_position, bin).await?;
     }
 
     if let Some(m) = metadata {
@@ -61,4 +63,20 @@ where
     write_chunks(writer, bin.chunks()).await?;
 
     Ok(())
+}
+
+fn first_record_start_position(index: &BinnedIndex, mut id: usize) -> bgzf::VirtualPosition {
+    let mut min_position = index.get(&id).copied().unwrap_or_default();
+
+    while let Some(pid) = parent_id(id)
+        && let Some(position) = index.get(&pid)
+    {
+        if *position < min_position {
+            min_position = *position;
+        }
+
+        id = pid;
+    }
+
+    min_position
 }